由于我早上太能熬夜，忘了测试自动启动和图标了
现在我修复了desktop 的图标 ，图标需要一个png/svg格式，jpg不支持，自动启动忘判断了

+++ showjcr.cpp
@@ -239,19 +239,20 @@ void ShowJCR::setAutoStart()
 					QString fullPath = dir.filePath("io.hitfyd.ShowJCR.desktop");
 					if (QFileInfo::exists(fullPath)) {
 							sourceFile=fullPath; // 找到文件，返回路径
+							break;
 					}
 			}
 		QString configDir = QStandardPaths::writableLocation(QStandardPaths::ConfigLocation);
-		QDir autostartDir(configDir + "autostart");
+		QDir autostartDir(configDir + "/autostart/");
 		if (!autostartDir.exists()) {
 			if (!autostartDir.mkpath(".")) {
 					return;
 				}
 		}
 		QString desktopFilePath = autostartDir.filePath("io.hitfyd.ShowJCR.desktop");
-		if (QFile::exists(desktopFilePath)){
+		if (QFile::exists(desktopFilePath)&&!autoStart){
 				QFile::remove(desktopFilePath);
-			}else{
+			}else if(autoStart&&!QFile::exists(desktopFilePath)){
 				QFile::copy(sourceFile, desktopFilePath);
 			}
 	#endif
